+++ AMaze/Player.cs
@@ -2,23 +2,19 @@
 
 internal class Player
 {
+	private readonly int[] keyCount = new int[KeyShapeExt.Count];
+
 	public double HitboxHalf { get; }
 	public double InteractionDist { get; }
 
 	public double X { get; set; }
 	public double Y { get; set; }
 	public double Rot { get; set; }
-	public int KeyCount { get; set; }
 
 	public Player(double hitboxSize, double interactionDist)
 	{
 		HitboxHalf = hitboxSize / 2.0;
 		InteractionDist = interactionDist;
-	}
-
-	private Geometry.Rect MakeHitbox(double x, double y)
-	{
-		return new Geometry.Rect { x1 = x - HitboxHalf, y1 = y - HitboxHalf, x2 = x + HitboxHalf, y2 = y + HitboxHalf };
 	}
 
 	public void Move(double step, double rotOffset, Span<Entities.IEntity> entities)
@@ -39,10 +35,14 @@ internal class Player
 	}
 	private static bool DoesCollide(Geometry.Rect hitbox, Span<Entities.IEntity> entities)
 	{
-		foreach (Entities.IEntity entity in entities)
-			if (entity.DoesCollide(hitbox))
+		for (int i = 0; i < entities.Length; i++)
+			if (entities[i].DoesCollide(hitbox))
 				return true;
 		return false;
+	}
+	private Geometry.Rect MakeHitbox(double x, double y)
+	{
+		return new Geometry.Rect { x1 = x - HitboxHalf, y1 = y - HitboxHalf, x2 = x + HitboxHalf, y2 = y + HitboxHalf };
 	}
 
 	public void Rotate(double step)
@@ -56,15 +56,43 @@ internal class Player
 
 	public void Interact(Span<Entities.IEntity> entities)
 	{
-		var seg = new Geometry.Seg {
-			x1 = X,
-			y1 = Y,
-			x2 = X + Math.Cos(Rot) * InteractionDist,
-			y2 = Y + Math.Sin(Rot) * InteractionDist,
+		var seg = MakeInteractionSeg();
+		for (int i = 0; i < entities.Length; i++)
+			if ((entities[i] is Entities.IInteractable interactable) && entities[i].Intersect(seg, out var _))
+				interactable.Interact();
+	}
+	public bool CanInteract(Span<Entities.IEntity> entities)
+	{
+		var seg = MakeInteractionSeg();
+		for (int i = 0; i < entities.Length; i++)
+			if ((entities[i] is Entities.IInteractable interactable) && entities[i].Intersect(seg, out var _)
+					&& interactable.CanInteract())
+				return true;
+		return false;
+	}
+	private Geometry.Seg MakeInteractionSeg()
+	{
+		return new Geometry.Seg {
+			x1 = X, x2 = X + Math.Cos(Rot) * InteractionDist,
+			y1 = Y, y2 = Y + Math.Sin(Rot) * InteractionDist,
 		};
-		foreach (Entities.IEntity entity in entities)
-			if (entity is Entities.IInteractable interactable)
-				if (entity.Intersect(seg, out var _))
-					interactable.Interact();
+	}
+
+	public int KeyCount(KeyShape shape)
+	{
+		return keyCount[(int)shape];
+	}
+	public void AddKey(KeyShape shape)
+	{
+		keyCount[(int)shape]++;
+	}
+	public bool UseKey(KeyShape shape)
+	{
+		if (keyCount[(int)shape] > 0)
+		{
+			keyCount[(int)shape]--;
+			return true;
+		}
+		return false;
 	}
 }